requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pip==24.0
python-dotenv==1.0.1
requests==2.32.2
SQLAlchemy==2.0.30
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
Zenora==0.0.3.post1
